Use RunningClient info to avoid deleting versions that are in use by running clients.

+++ SS14.Launcher/Models/Updater.cs
@@ -327,6 +327,9 @@ public sealed partial class Updater : ReactiveObject
 
         var versions = con.Query<ContentVersion>("SELECT * FROM ContentVersion ORDER BY LastUsed DESC").ToArray();
 
+        // NOTE: GetRunningClientVersions may modify DB, best to let it commit for cleanup.
+        var usedVersions = ContentManager.GetRunningClientVersions(con);
+
         var forkCounts = versions.Select(x => x.ForkId).Distinct().ToDictionary(x => x, _ => 0);
 
         var totalCount = 0;
@@ -342,9 +345,19 @@ public sealed partial class Updater : ReactiveObject
             }
             else
             {
-                Log.Debug("Culling version {ForkId}/{ForkVersion}", version.ForkId, version.ForkVersion);
-                con.Execute("DELETE FROM ContentVersion WHERE Id = @Id", new { version.Id });
-                anythingRemoved = true;
+                if (usedVersions.Contains(version.Id))
+                {
+                    Log.Debug(
+                        "Not culling version {ForkId}/{ForkVersion}: in use by running client",
+                        version.ForkId,
+                        version.ForkVersion);
+                }
+                else
+                {
+                    Log.Debug("Culling version {ForkId}/{ForkVersion}", version.ForkId, version.ForkVersion);
+                    con.Execute("DELETE FROM ContentVersion WHERE Id = @Id", new { version.Id });
+                    anythingRemoved = true;
+                }
             }
         }
